+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.6.5] - 2018-08-22
+### Changed
+- Update specification filter to get slug from link.
+
+### Fixed
+- Error on normalize function of the product summary.
+
 ## [0.6.4] - 2018-08-08
 ### Added
 - PriceRange filter.